Initial Transformer Setup
Added Transformer setup that allows us to incrementally port our transpile functions to transform functions.

We have to prot all transpile functions that transpile Nodes that are not allowed in the 2nd AST.

Once thats done we should restrucutre the whole project, because right now the transform process is just hacked before the transpilation.